move Emoji type to not pull in DOM things in tests
