+++ apps/zbugs/src/pages/issue/comment.tsx
@@ -4,10 +4,11 @@ import {memo, useState} from 'react';
 import {Button} from '../../components/button.js';
 import {CanEdit} from '../../components/can-edit.js';
 import {Confirm} from '../../components/confirm.js';
-import {EmojiPanel, type Emoji} from '../../components/emoji-panel.js';
+import {EmojiPanel} from '../../components/emoji-panel.js';
 import {Link} from '../../components/link.js';
 import Markdown from '../../components/markdown.js';
 import RelativeTime from '../../components/relative-time.js';
+import {type Emoji} from '../../emoji-utils.js';
 import {useHash} from '../../hooks/use-hash.js';
 import {useLogin} from '../../hooks/use-login.js';
 import {useZero} from '../../hooks/use-zero.js';
